+++ src/main/java/com/ddobang/backend/global/exception/oauth2/OAuth2Exception.java
@@ -1,0 +1,9 @@
+package com.ddobang.backend.global.exception.oauth2;
+
+import com.ddobang.backend.global.exception.ServiceException;
+
+public class OAuth2Exception extends ServiceException {
+	public OAuth2Exception(OAuth2ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
